# openapi/nctr.yaml
definitions:
...
  DepositData:
    properties:
      txHash:
        example: "0xbc8c24b7f1c2e280558c43fcfcc5ebe140619c1ebe70d035aa470b69778ec3c9"
        type: string


    type: object
info:
  contact: { }
  description: nctr debug api
  title: nctr project
  version: "1.0"
paths:
...
  /harvest/deposit:
    get:
      consumes:
        - 'Accept: */*'
      produces:
        - application/json
      parameters:
        - description: deposit amount
          in: query
          name: amount
          type: number
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/DepositData'
      summary: harvest deposit
swagger: "2.0"